Задание 5: Разработать аспект @Metric:
- Измеряет время работы метода

- Если время работы метода выше лимита, установленного в конфиге:

- Отправить сообщение в топик Кафки service_logs с информацией о сигнатуре метода, тип - WARNING, временем работы, входными параметрами, если таковые имеются

- Лимит выполнения устанавливается в application.yml

Разработать аспект @Cached

- Кэширует записи из БД

- Перед выполнением запроса в БД кэш должен проверяться на наличие нужной записи

- Если таковой нет - сделать запрос

- В качестве ключа - primary key, если хранилище индивидуально под сущность, или hashCode объекта, если кэш общий

- Время хранения записи в кэше задать в application.yml

- По истечению времени запись должна удаляться из кэша

Так же чуть рефакторинг

+++ first-homework/account-processing/src/main/java/com/t1/marselmkh/repository/TransactionRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.time.LocalDateTime;
 
 @Repository
-public interface TransactionRepository extends CrudRepository<Transaction,Long> {
-    int countByCardIdAndTimestampBetween(Long cardId, LocalDateTime from, LocalDateTime now);
+public interface TransactionRepository extends CrudRepository<Transaction, Long> {
+    Long countByCardIdAndTimestampBetween(Long cardId, LocalDateTime from, LocalDateTime now);
 }
